fixed last detectable bug + removal of redundant codes + optimization
Removed unused codes (calibration stuff, fastest path algo, and exploration)

fixed algo to rotate and sense obstacles of the direction it wants to check if free to move to. IE, if want check northFree(), robot rotate and sense 3 cells of north first beforehand if any of the 3 cells not explored. CAN BE IMPROVED FURTHER (PREFERENCE TO ROTATE 1 STEP LEFT THAN 2 STEPS RIGHT)

optimized bottom-up looping of fastestpath to top-down